feat: 템플릿 메소드 패턴 적용, 도메인 의존도 제거
