+++ jdbc/src/main/java/nextstep/jdbc/exception/JdbcException.java
@@ -1,0 +1,8 @@
+package nextstep.jdbc.exception;
+
+public class JdbcException extends RuntimeException {
+
+    public JdbcException(String message) {
+        super(message);
+    }
+}
